// src/lib/infrastructure/api/base.ts
/* tslint:disable */
/* eslint-disable */
/**
 * λORM Service
 * service of lambda orm
 *
 * The version of the OpenAPI document: 0.3.0
 * Contact: flaviolrita@hotmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { Configuration } from "../../domain/configuration";
// Some imports not used depending on template conditions
// @ts-ignore
import globalAxios, { AxiosPromise, AxiosInstance, AxiosRequestConfig } from 'axios';
/**
 *
 * @export
 */
export const COLLECTION_FORMATS = {
    csv: ",",
    ssv: " ",
    tsv: "\t",
    pipes: "|",
};

/**
 *
 * @export
 * @interface RequestArgs
 */
export interface RequestArgs {
    url: string;
    options: AxiosRequestConfig;
}

/**
 *
 * @export
 * @class BaseAPI
 */
export class BaseAPI {
    protected configuration: Configuration
    protected basePath:string | undefined

    constructor(configuration: Configuration,protected axios: AxiosInstance = globalAxios) {        
        this.configuration = configuration;
        this.basePath = configuration.basePath || 'http://0.0.0.0/9291'
    }
};

/**
 *
 * @export
 * @class RequiredError
 * @extends {Error}
 */
export class RequiredError extends Error {
    name: "RequiredError" = "RequiredError";
    constructor(public field: string, msg?: string) {
        super(msg);
    }
}